GSC demo: ease the motion of the windvane when changing value
looking to see if I can also animate the motion of the triangle moving,
so we can see the colours sweep across sensors when changing value. may
be trickier to do as we're doing a manual transform rather than a css
transform.

look into whether the path created by leaflet can be animated to
transform in the same way the windvane does.