GSC demo: fixes. popover. extract class to organise code
Created a class for an emission source as was ending up coupling a bunch
of stuff together in multiple places (such as which emission source for
placing a marker...vs emission source just to show popover). Made it
tricky to reason about where all changes would occur.

This lets us centralise the logic for popovers and pollution markers
etc.