fix bug in minesweeper. now only check hidden sensors
We had a bug where you could click near an already revealed sensor and
it would increment the 'found' counter, thus falsely ending the game
early.

Added a 'visible' flag to the pollutant class for use in this game mode,
and now only check the hidden pollutants for proximity to click.