better scenario css. better random emission generation.
use variables in css.

use rules PER SCENARIO, rather than trying to add and remove styles and
do a bunch of hacks. This makes hiding and showing sections easier, as I
can just hide entire sections and then only toggle display using CSS and
a css class add.

generate random emissions on a grid, to ensure a minimum distance.
Probably want to convert this to be a fairly wide grid, but then with a
bit of wobble afterwards so they don't look so square on a grid.
However, this should hopefully let me get emission sources that are not
directly on top of each other.