s3: multiple emissions sources. all: better styling for ipad
styling was a bit spread out leaving not much content visible on ipad.
tweaked a bunch of spacing to try and get it appearing a bit better.
still need to think better how to convey information.

also need to improve how the logo fits.

I broke scenario3 when I implemented the Scenario class, as I didn't
include a callback to set N_POLLUTANTS and CYCLE_POLLUTANTS. Now, 3
pollutants, and oldest will be cycled out on click.